Multiplanar: Use RasterInterface in VideoResourceUpdater
This change updates VideoResourceUpdater to use RasterInterface instead
of Gles2Interface. The change means that it now goes through
RasterInterface which calls RasterImplementationGLES which in turn
still uses gles interface. This in turn updates to use
CopySharedImage and WritePixels apis instead of CopySubTexture and
TexSubImage2D directly.

It also updates RasterImplementationGLES util methods to get the
right GL data format and data type to support 10/16 bit formats.

Bug: 1465939
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4827410
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1194314}

+++ media/renderers/video_resource_updater.cc
@@ -43,6 +43,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "gpu/command_buffer/common/shared_image_trace_utils.h"
 #include "gpu/command_buffer/common/shared_image_usage.h"
 #include "media/base/format_utils.h"
+#include "media/base/media_switches.h"
 #include "media/base/wait_and_replace_sync_token_client.h"
 #include "media/renderers/paint_canvas_video_renderer.h"
 #include "media/renderers/resource_sync_token_client.h"
@@ -281,12 +282,6 @@ VideoFrameResourceType ExternalResourceTypeForHardwarePlanes(
   return VideoFrameResourceType::NONE;
 }
 
-// Sync tokens passed downstream to the compositor can be unverified.
-void GenerateCompositorSyncToken(gpu::gles2::GLES2Interface* gl,
-                                 gpu::SyncToken* sync_token) {
-  gl->GenUnverifiedSyncTokenCHROMIUM(sync_token->GetData());
-}
-
 // For frames that we receive in software format, determine the dimensions of
 // each plane in the frame.
 gfx::Size SoftwarePlaneDimension(VideoFrame* input_frame,
@@ -346,6 +341,12 @@ bool HasCompatibleFormat(VideoPixelFormat input_format,
   if (input_format == PIXEL_FORMAT_ARGB)
     return output_format == viz::SinglePlaneFormat::kBGRA_8888;
   return false;
+}
+
+// Returns if kRasterInterfaceInVideoResourceUpdater is enabled
+bool CanUseRasterInterface() {
+  return base::FeatureList::IsEnabled(
+      media::kRasterInterfaceInVideoResourceUpdater);
 }
 
 class CopyingSyncTokenClient : public VideoFrame::SyncTokenClient {
@@ -552,7 +553,7 @@ class VideoResourceUpdater::HardwarePlaneResource
         format, size, color_space, kTopLeft_GrSurfaceOrigin,
         kPremul_SkAlphaType, shared_image_usage, "VideoResourceUpdater",
         gpu::kNullSurfaceHandle);
-    ContextGL()->WaitSyncTokenCHROMIUM(
+    InterfaceBase()->WaitSyncTokenCHROMIUM(
         sii->GenUnverifiedSyncToken().GetConstData());
   }
 
@@ -561,7 +562,7 @@ class VideoResourceUpdater::HardwarePlaneResource
 
   ~HardwarePlaneResource() override {
     gpu::SyncToken sync_token;
-    ContextGL()->GenUnverifiedSyncTokenCHROMIUM(sync_token.GetData());
+    InterfaceBase()->GenUnverifiedSyncTokenCHROMIUM(sync_token.GetData());
     SharedImageInterface()->DestroySharedImage(sync_token, mailbox_);
   }
 
@@ -581,6 +582,18 @@ class VideoResourceUpdater::HardwarePlaneResource
     auto* gl = context_provider_->ContextGL();
     DCHECK(gl);
     return gl;
+  }
+
+  gpu::raster::RasterInterface* RasterInterface() {
+    auto* ri = context_provider_->RasterInterface();
+    CHECK(ri);
+    return ri;
+  }
+
+  gpu::InterfaceBase* InterfaceBase() {
+    return CanUseRasterInterface()
+               ? static_cast<gpu::InterfaceBase*>(RasterInterface())
+               : static_cast<gpu::InterfaceBase*>(ContextGL());
   }
 
   const raw_ptr<viz::RasterContextProvider> context_provider_;
@@ -931,27 +944,42 @@ void VideoResourceUpdater::CopyHardwarePlane(
   DCHECK_EQ(hardware_resource->texture_target(),
             static_cast<GLenum>(GL_TEXTURE_2D));
 
-  auto* gl = ContextGL();
-  gl->WaitSyncTokenCHROMIUM(mailbox_holder.sync_token.GetConstData());
+  if (CanUseRasterInterface()) {
+    auto* ri = RasterInterface();
+    ri->WaitSyncTokenCHROMIUM(mailbox_holder.sync_token.GetConstData());
 
-  // This is only used on Android where all video mailboxes already use shared
-  // images.
-  DCHECK(mailbox_holder.mailbox.IsSharedImage());
-
-  // TODO(vikassoni): Use raster interface instead of gl interface eventually.
-  GLuint src_texture_id =
-      gl->CreateAndTexStorage2DSharedImageCHROMIUM(mailbox_holder.mailbox.name);
-  gl->BeginSharedImageAccessDirectCHROMIUM(
-      src_texture_id, GL_SHARED_IMAGE_ACCESS_MODE_READ_CHROMIUM);
-  {
-    HardwarePlaneResource::ScopedTexture scope(gl, hardware_resource);
-    gl->CopySubTextureCHROMIUM(
-        src_texture_id, 0, GL_TEXTURE_2D, scope.texture_id(), 0, 0, 0, 0, 0,
+    // This is only used on Android where all video mailboxes already use shared
+    // images.
+    CHECK(mailbox_holder.mailbox.IsSharedImage());
+    ri->CopySharedImage(
+        mailbox_holder.mailbox, hardware_resource->mailbox(), GL_TEXTURE_2D,
+        /*xoffset=*/0, /*yoffset=*/0, /*x=*/0, /*y=*/0,
         output_plane_resource_size.width(), output_plane_resource_size.height(),
-        false, false, false);
+        /*unpack_flip_y=*/false, /*unpack_premultiply_alpha=*/false);
+  } else {
+    auto* gl = ContextGL();
+    gl->WaitSyncTokenCHROMIUM(mailbox_holder.sync_token.GetConstData());
+
+    // This is only used on Android where all video mailboxes already use shared
+    // images.
+    DCHECK(mailbox_holder.mailbox.IsSharedImage());
+    GLuint src_texture_id = gl->CreateAndTexStorage2DSharedImageCHROMIUM(
+        mailbox_holder.mailbox.name);
+    gl->BeginSharedImageAccessDirectCHROMIUM(
+        src_texture_id, GL_SHARED_IMAGE_ACCESS_MODE_READ_CHROMIUM);
+    {
+      HardwarePlaneResource::ScopedTexture scope(gl, hardware_resource);
+      gl->CopySubTextureCHROMIUM(
+          src_texture_id, /*source_level=*/0, GL_TEXTURE_2D, scope.texture_id(),
+          /*dest_level=*/0, /*xoffset=*/0, /*yoffset=*/0, /*x=*/0, /*y=*/0,
+          output_plane_resource_size.width(),
+          output_plane_resource_size.height(),
+          /*unpack_flip_y=*/false, /*unpack_premultiply_alpha=*/false,
+          /*unpack_unmultiply_alpha=*/false);
+    }
+    gl->EndSharedImageAccessDirectCHROMIUM(src_texture_id);
+    gl->DeleteTextures(1, &src_texture_id);
   }
-  gl->EndSharedImageAccessDirectCHROMIUM(src_texture_id);
-  gl->DeleteTextures(1, &src_texture_id);
 
   // Wait (if the existing token isn't null) and replace it with a new one.
   //
@@ -959,7 +987,7 @@ void VideoResourceUpdater::CopyHardwarePlane(
   // here since this code isn't tuned for multiple planes; it should only update
   // the release token once.
   DCHECK_EQ(video_frame->NumTextures(), 1u);
-  WaitAndReplaceSyncTokenClient client(gl);
+  WaitAndReplaceSyncTokenClient client(InterfaceBase());
   gpu::SyncToken sync_token = video_frame->UpdateReleaseSyncToken(&client);
 
   auto transferable_resource = viz::TransferableResource::MakeGpu(
@@ -1294,14 +1322,25 @@ VideoFrameExternalResources VideoResourceUpdater::CreateForSoftwarePlanes(
         }
 
         // Copy pixels into texture.
-        auto* gl = ContextGL();
-        {
+        if (CanUseRasterInterface()) {
+          auto* ri = RasterInterface();
+          auto color_type = viz::ToClosestSkColorType(
+              /*gpu_compositing=*/true, output_si_format, /*plane_index=*/0);
+          SkImageInfo info =
+              SkImageInfo::Make(plane_size.width(), plane_size.height(),
+                                color_type, kPremul_SkAlphaType);
+          SkPixmap pixmap = SkPixmap(info, source_pixels, bytes_per_row);
+          ri->WritePixels(hardware_resource->mailbox(), /*dst_x_offset=*/0,
+                          /*dst_y_offset=*/0, /*dst_plane_index=*/0,
+                          hardware_resource->texture_target(), pixmap);
+        } else {
+          auto* gl = ContextGL();
           HardwarePlaneResource::ScopedTexture scope(gl, hardware_resource);
           gl->BindTexture(hardware_resource->texture_target(),
                           scope.texture_id());
           gl->TexSubImage2D(
-              hardware_resource->texture_target(), 0, 0, 0, plane_size.width(),
-              plane_size.height(),
+              hardware_resource->texture_target(), /*level=*/0, /*xoffset=*/0,
+              /*yoffset=*/0, plane_size.width(), plane_size.height(),
               viz::SharedImageFormatRestrictedSinglePlaneUtils::ToGLDataFormat(
                   output_si_format),
               viz::SharedImageFormatRestrictedSinglePlaneUtils::ToGLDataType(
@@ -1323,8 +1362,7 @@ VideoFrameExternalResources VideoResourceUpdater::CreateForSoftwarePlanes(
       HardwarePlaneResource* hardware_resource = plane_resource->AsHardware();
       external_resources.type = VideoFrameResourceType::RGBA;
       gpu::SyncToken sync_token;
-      auto* gl = ContextGL();
-      GenerateCompositorSyncToken(gl, &sync_token);
+      InterfaceBase()->GenUnverifiedSyncTokenCHROMIUM(sync_token.GetData());
       transferable_resource = viz::TransferableResource::MakeGpu(
           hardware_resource->mailbox(), hardware_resource->texture_target(),
           sync_token, hardware_resource->resource_size(), output_si_format,
@@ -1358,7 +1396,6 @@ VideoFrameExternalResources VideoResourceUpdater::CreateForSoftwarePlanes(
     external_resources.offset = 0;
   }
 
-  auto* gl = ContextGL();
   // We need to transfer data from |video_frame| to the plane resources.
   for (size_t i = 0; i < plane_resources.size(); ++i) {
     HardwarePlaneResource* plane_resource = plane_resources[i]->AsHardware();
@@ -1405,6 +1442,7 @@ VideoFrameExternalResources VideoResourceUpdater::CreateForSoftwarePlanes(
     GLuint unpack_alignment = kDefaultUnpackAlignment;
 
     const uint8_t* pixels;
+    int pixels_stride_in_bytes;
 
     if (!needs_conversion) {
       // Stride adaptation is needed if source and destination strides are
@@ -1425,6 +1463,7 @@ VideoFrameExternalResources VideoResourceUpdater::CreateForSoftwarePlanes(
           unpack_alignment = bytes_per_element;
       }
       pixels = video_frame->data(i);
+      pixels_stride_in_bytes = video_stride_bytes;
     } else {
       // Avoid malloc for each frame/plane if possible.
       const size_t needed_size =
@@ -1457,11 +1496,24 @@ VideoFrameExternalResources VideoResourceUpdater::CreateForSoftwarePlanes(
       }
 
       pixels = upload_pixels_.get();
+      pixels_stride_in_bytes = upload_image_stride;
     }
 
     // Copy pixels into texture. TexSubImage2D() is applicable because
     // |yuv_si_format| is LUMINANCE_F16, R16_EXT, LUMINANCE_8 or RED_8.
-    {
+    if (CanUseRasterInterface()) {
+      auto* ri = RasterInterface();
+      auto color_type = viz::ToClosestSkColorType(
+          /*gpu_compositing=*/true, plane_si_format, /*plane_index=*/0);
+      SkImageInfo info = SkImageInfo::Make(resource_size_pixels.width(),
+                                           resource_size_pixels.height(),
+                                           color_type, kPremul_SkAlphaType);
+      SkPixmap pixmap = SkPixmap(info, pixels, pixels_stride_in_bytes);
+      ri->WritePixels(plane_resource->mailbox(), /*dst_x_offset=*/0,
+                      /*dst_y_offset=*/0, /*dst_plane_index=*/0,
+                      plane_resource->texture_target(), pixmap);
+    } else {
+      auto* gl = ContextGL();
       HardwarePlaneResource::ScopedTexture scope(gl, plane_resource);
 
       gl->BindTexture(plane_resource->texture_target(), scope.texture_id());
@@ -1469,8 +1521,9 @@ VideoFrameExternalResources VideoResourceUpdater::CreateForSoftwarePlanes(
       gl->PixelStorei(GL_UNPACK_ROW_LENGTH, unpack_row_length);
       gl->PixelStorei(GL_UNPACK_ALIGNMENT, unpack_alignment);
       gl->TexSubImage2D(
-          plane_resource->texture_target(), 0, 0, 0,
-          resource_size_pixels.width(), resource_size_pixels.height(),
+          plane_resource->texture_target(), /*level=*/0, /*xoffset=*/0,
+          /*yoffset=*/0, resource_size_pixels.width(),
+          resource_size_pixels.height(),
           viz::SharedImageFormatRestrictedSinglePlaneUtils::ToGLDataFormat(
               plane_si_format),
           viz::SharedImageFormatRestrictedSinglePlaneUtils::ToGLDataType(
@@ -1485,7 +1538,7 @@ VideoFrameExternalResources VideoResourceUpdater::CreateForSoftwarePlanes(
 
   // Set the sync token otherwise resource is assumed to be synchronized.
   gpu::SyncToken sync_token;
-  GenerateCompositorSyncToken(gl, &sync_token);
+  InterfaceBase()->GenUnverifiedSyncTokenCHROMIUM(sync_token.GetData());
 
   for (size_t i = 0; i < plane_resources.size(); ++i) {
     HardwarePlaneResource* plane_resource = plane_resources[i]->AsHardware();
@@ -1512,6 +1565,18 @@ gpu::gles2::GLES2Interface* VideoResourceUpdater::ContextGL() {
   return gl;
 }
 
+gpu::raster::RasterInterface* VideoResourceUpdater::RasterInterface() {
+  auto* ri = context_provider_->RasterInterface();
+  CHECK(ri);
+  return ri;
+}
+
+gpu::InterfaceBase* VideoResourceUpdater::InterfaceBase() {
+  return CanUseRasterInterface()
+             ? static_cast<gpu::InterfaceBase*>(RasterInterface())
+             : static_cast<gpu::InterfaceBase*>(ContextGL());
+}
+
 void VideoResourceUpdater::ReturnTexture(
     scoped_refptr<VideoFrame> video_frame,
     const gpu::SyncToken& original_release_token,
@@ -1528,7 +1593,7 @@ void VideoResourceUpdater::ReturnTexture(
     return;
   }
 
-  ResourceSyncTokenClient client(ContextGL(), original_release_token,
+  ResourceSyncTokenClient client(InterfaceBase(), original_release_token,
                                  new_release_token);
   video_frame->UpdateReleaseSyncToken(&client);
 }
@@ -1542,7 +1607,7 @@ void VideoResourceUpdater::RecycleResource(uint32_t plane_resource_id,
     return;
 
   if (context_provider_ && sync_token.HasData()) {
-    ContextGL()->WaitSyncTokenCHROMIUM(sync_token.GetConstData());
+    InterfaceBase()->WaitSyncTokenCHROMIUM(sync_token.GetConstData());
   }
 
   if (lost_resource) {
